GitHub #159: Restart DSP backend on error (first pass)
Still need to add throttling so that we don't get stuck in a loop if the
patch load has a problem.

+++ mfp/processor.py
@@ -402,6 +402,8 @@ class Processor (object):
         from .mfp_app import MFPApp
         from .patch import Patch
 
+        print "Processor.delete: enter"
+
         if hasattr(self, "patch") and self.patch is not None:
             self.patch.unbind(self.name, self.scope)
             self.patch.remove(self)
@@ -438,7 +440,8 @@ class Processor (object):
             self.midi_cbid = None 
 
         if hasattr(self, "dsp_obj") and self.dsp_obj is not None:
-            self.dsp_obj.delete()
+            if self.patch.context is not None: 
+                self.dsp_obj.delete()
             self.dsp_obj = None
 
         MFPApp().forget(self)
